fix codex

+++ interfaces/inner_api/common/taginfo_parcelable.cpp
@@ -64,7 +64,7 @@ TagInfoParcelable *TagInfoParcelable::Unmarshalling(Parcel &parcel)
 
     int32_t extraLen = 0;
     parcel.ReadInt32(extraLen);
-    if (extraLen >= MAX_TECH_LIST_NUM) {
+    if (extraLen >= static_cast<int32_t>(MAX_TECH_LIST_NUM)) {
         return nullptr;
     }
     std::vector<AppExecFwk::PacMap> tagTechExtrasData;
